+++ assets/js/theme/account.js
@@ -1,10 +1,12 @@
 import PageManager from './page-manager';
+import _ from 'lodash';
 import $ from 'jquery';
 import nod from './common/nod';
 import Wishlist from './wishlist';
 import validation from './common/form-validation';
 import stateCountry from './common/state-country';
 import { classifyForm, Validators, insertStateHiddenField } from './common/form-utils';
+import { creditCardType, storeInstrument, Validators as CCValidators, Formatters as CCFormatters } from './common/payment-method';
 import swal from 'sweetalert2';
 
 export default class Account extends PageManager {
@@ -20,6 +22,7 @@ export default class Account extends PageManager {
         const $addressForm = classifyForm('form[data-address-form]');
         const $inboxForm = classifyForm('form[data-inbox-form]');
         const $accountReturnForm = classifyForm('[data-account-return-form]');
+        const $paymentMethodForm = classifyForm('form[data-payment-method-form]');
         const $reorderForm = classifyForm('[data-account-reorder-form]');
         const $invoiceButton = $('[data-print-invoice]');
 
@@ -62,11 +65,16 @@ export default class Account extends PageManager {
             this.initAccountReturnFormValidation($accountReturnForm);
         }
 
+        if ($paymentMethodForm.length) {
+            this.initPaymentMethodFormValidation($paymentMethodForm);
+        }
+
         if ($reorderForm.length) {
             this.initReorderForm($reorderForm);
         }
 
         this.bindDeleteAddress();
+        this.bindDeletePaymentMethod();
     }
 
     /**
@@ -75,6 +83,16 @@ export default class Account extends PageManager {
     bindDeleteAddress() {
         $('[data-delete-address]').on('submit', event => {
             const message = $(event.currentTarget).data('deleteAddress');
+
+            if (!window.confirm(message)) {
+                event.preventDefault();
+            }
+        });
+    }
+
+    bindDeletePaymentMethod() {
+        $('[data-delete-payment-method]').on('submit', event => {
+            const message = $(event.currentTarget).data('deletePaymentMethod');
 
             if (!window.confirm(message)) {
                 event.preventDefault();
@@ -187,6 +205,106 @@ export default class Account extends PageManager {
             });
 
             return event.preventDefault();
+        });
+    }
+
+    initPaymentMethodFormValidation($paymentMethodForm) {
+        // Inject validations into form fields before validation runs
+        $paymentMethodForm.find('#first_name.form-field').attr('data-validation', `{ "type": "singleline", "label": "${this.context.firstNameLabel}", "required": true, "maxlength": 0 }`);
+        $paymentMethodForm.find('#last_name.form-field').attr('data-validation', `{ "type": "singleline", "label": "${this.context.lastNameLabel}", "required": true, "maxlength": 0 }`);
+        $paymentMethodForm.find('#company.form-field').attr('data-validation', `{ "type": "singleline", "label": "${this.context.companyLabel}", "required": false, "maxlength": 0 }`);
+        $paymentMethodForm.find('#phone.form-field').attr('data-validation', `{ "type": "singleline", "label": "${this.context.phoneLabel}", "required": false, "maxlength": 0 }`);
+        $paymentMethodForm.find('#address1.form-field').attr('data-validation', `{ "type": "singleline", "label": "${this.context.address1Label}", "required": true, "maxlength": 0 }`);
+        $paymentMethodForm.find('#address2.form-field').attr('data-validation', `{ "type": "singleline", "label": "${this.context.address2Label}", "required": false, "maxlength": 0 }`);
+        $paymentMethodForm.find('#city.form-field').attr('data-validation', `{ "type": "singleline", "label": "${this.context.cityLabel}", "required": true, "maxlength": 0 }`);
+        $paymentMethodForm.find('#country.form-field').attr('data-validation', `{ "type": "singleselect", "label": "${this.context.countryLabel}", "required": true, prefix: "${this.context.chooseCountryLabel}" }`);
+        $paymentMethodForm.find('#state.form-field').attr('data-validation', `{ "type": "singleline", "label": "${this.context.stateLabel}", "required": true, "maxlength": 0 }`);
+        $paymentMethodForm.find('#postal_code.form-field').attr('data-validation', `{ "type": "singleline", "label": "${this.context.postalCodeLabel}", "required": true, "maxlength": 0 }`);
+
+        const validationModel = validation($paymentMethodForm);
+        const paymentMethodSelector = 'form[data-payment-method-form]';
+        const paymentMethodValidator = nod({ submit: `${paymentMethodSelector} input[type="submit"]` });
+        const $stateElement = $(`${paymentMethodSelector} [data-field-type="State"]`);
+
+        let $last;
+        // Requests the states for a country with AJAX
+        stateCountry($stateElement, this.context, (err, field) => {
+            if (err) {
+                throw new Error(err);
+            }
+
+            const $field = $(field);
+
+            if (paymentMethodValidator.getStatus($stateElement) !== 'undefined') {
+                paymentMethodValidator.remove($stateElement);
+            }
+
+            if ($last) {
+                paymentMethodValidator.remove($last);
+            }
+
+            if ($field.is('select')) {
+                $last = field;
+                Validators.setStateCountryValidation(paymentMethodValidator, field);
+            } else {
+                Validators.cleanUpStateValidation(field);
+            }
+        });
+
+        // Use credit card number input listener to highlight credit card type
+        $(`${paymentMethodSelector} input[name="credit_card_number"]`).on('keyup', ({ target }) => {
+            const cardType = creditCardType(target.value);
+            if (cardType) {
+                $(`${paymentMethodSelector} img[alt="${cardType}"`).siblings().css('opacity', '.2');
+            } else {
+                $(`${paymentMethodSelector} img`).css('opacity', '1');
+            }
+        });
+
+        // Set of credit card validation
+        CCValidators.setCreditCardNumberValidation(paymentMethodValidator, `${paymentMethodSelector} input[name="credit_card_number"]`, this.context.creditCardNumber);
+        CCValidators.setExpirationValidation(paymentMethodValidator, `${paymentMethodSelector} input[name="expiration"]`, this.context.expiration);
+        CCValidators.setNameOnCardValidation(paymentMethodValidator, `${paymentMethodSelector} input[name="name_on_card"]`, this.context.nameOnCard);
+        CCValidators.setCvvValidation(paymentMethodValidator, `${paymentMethodSelector} input[name="cvv"]`, this.context.cvv);
+
+        // Set of credit card format
+        CCFormatters.setCreditCardNumberFormat(`${paymentMethodSelector} input[name="credit_card_number"]`);
+        CCFormatters.setExpirationFormat(`${paymentMethodSelector} input[name="expiration"`);
+
+        // Billing address validation
+        paymentMethodValidator.add(validationModel);
+
+        $paymentMethodForm.on('submit', event => {
+            event.preventDefault();
+            // Perform final form validation
+            paymentMethodValidator.performCheck();
+            if (paymentMethodValidator.areAll('valid')) {
+                // Serialize form data and reduce it to object
+                const data = _.reduce($paymentMethodForm.serializeArray(), (obj, item) => {
+                    const refObj = obj;
+                    refObj[item.name] = item.value;
+                    return refObj;
+                }, {});
+
+                // Assign country and state code
+                const country = _.find(this.context.countries, ({ value }) => value === data.country);
+                const state = country && _.find(country.states, ({ value }) => value === data.state);
+                data.country_code = country ? country.code : data.country;
+                data.state_or_province_code = state ? state.code : data.state;
+
+                // Default Instrument
+                data.default_instrument = !!data.default_instrument;
+
+                // Store credit card
+                storeInstrument(this.context, data, () => {
+                    window.location.href = this.context.paymentMethodsUrl;
+                }, () => {
+                    swal({
+                        text: this.context.generic_error,
+                        type: 'error',
+                    });
+                });
+            }
         });
     }